Fix case where user tries to rename account to the current name.
Prior to this patch, the changing of the username to the same
username fails, because the worker here expects that no doc
exists in _users with the target name.

This patch addresses this by releting the `$newUsername`
property from the user doc and otherwise ignores the request.

The user-intent is already fulfilled.

This could probably also caught in the frontend.